requirements.txt
paho-mqtt=1.5.0
solaredge_modbus=0.2.0